+++ aeron-quick-base/src/main/java/org/jetc/aeron/quick/messaging/fragment_handling/concurrent/disruptor/DisruptorExecutorPool.java
@@ -5,8 +5,8 @@ import com.lmax.disruptor.RingBuffer;
 import com.lmax.disruptor.WaitStrategy;
 import com.lmax.disruptor.dsl.Disruptor;
 import com.lmax.disruptor.dsl.ProducerType;
-import org.jetc.aeron.quick.deferred_exec.ExecutionRequestEvent;
-import org.jetc.aeron.quick.deferred_exec.MethodExecutor;
+import org.jetc.aeron.quick.utils.events.ExecutionRequestEvent;
+import org.jetc.aeron.quick.utils.events.MethodExecutor;
 import org.jetc.aeron.quick.messaging.AeronBindingHolder;
 import org.jetc.aeron.quick.messaging.AeronConcurrentBinding;
 import org.jetc.aeron.quick.messaging.fragment_handling.concurrent.OperationExecutorPool;
