+++ aeron-quick-base/src/main/java/org/jetc/aeron/quick/messaging/publication/PublicationOfferingStrategy.java
@@ -2,7 +2,7 @@ package org.jetc.aeron.quick.messaging.publication;
 
 import io.aeron.Publication;
 import org.agrona.MutableDirectBuffer;
-import org.jetc.aeron.quick.exception.PublicationOfferFailedException;
+import org.jetc.aeron.quick.messaging.publication.exception.PublicationOfferFailedException;
 import java.util.function.Consumer;
 
 public interface PublicationOfferingStrategy {
